+++ src/server/api/trpc.ts
@@ -1,7 +1,7 @@
 import { TRPCError } from '@trpc/server'
 import { type PrismaClient } from '@prisma/client'
 import { getServerSession } from 'next-auth'
-import { authOptions } from '@/app/api/auth/[...nextauth]/route'
+import { authOptions } from '@/lib/auth'
 import { Enum } from '@/enums'
 import { middleware, procedure, router } from '@/server/trpc'
 import { ERROR_MESSAGES } from '@/constants/trpc'
@@ -11,8 +11,8 @@ import { Context } from '@/server/context'
 /**
  * 트랜잭션 미들웨어
  */
-export const transactionMiddleware = middleware(async ({ ctx, next }) => {
-	return await ctx.prisma.$transaction(async (tx) => {
+export const transactionMiddleware = middleware(({ ctx, next }) => {
+	return ctx.prisma.$transaction((tx) => {
 		return next({
 			ctx: {
 				...ctx,
